feat: post-incident intelligence (AI step 4) + durationMs overflow fix
Three explicit, user-initiated AI features closing the incident knowledge
loop — drafts and suggestions only, never auto-actions:

- Resolution draft: a "Bozza AI dalle attività" button in the transition
  dialog generates resolution/root-cause notes from the incident's REAL
  timeline (comments, workflow step executions, affected CIs). Honest by
  prompt: where the evidence is missing it says so instead of inventing.
  Verified live: 585-char draft correctly stating the cause was undocumented.
- Problem candidates: on the Problems page, one click clusters non-closed
  incidents by embedding similarity (union-find over vector-index neighbors,
  threshold 0.72, min cluster 3) and has Claude name each cluster's probable
  root cause with a motivation; the panel shows the evidence as clickable
  incident chips. Suggestion only — creating the Problem stays manual.
- KB draft from resolved incident: a button next to Export PDF (resolved/
  closed only) generates a structured article (Sintomo/Causa/Soluzione/
  Verifica) and creates it through the standard KB path in its initial
  draft state for review. Verified live: sensible title, category and tags
  extracted from the actual resolution comments.

postIncidentService uses the SDK with adaptive thinking, cached system
prompts, effort low, and structured outputs where shape matters; all
GraphQL surfaces are tenant-scoped; missing API key and refusals surface
as explicit errors.

Also fixes a real bug found while testing: WorkflowStepExecution.durationMs
(and the discovery twin) was GraphQL Int — any workflow step older than
~24.8 days overflowed 32 bits and broke the whole incident detail page with
an INTERNAL_SERVER_ERROR. Now Float.

e2e: postincident.spec.ts covers all three features with the explicit-
outcome contract; full suite 13/13 green on the deployed stack.

+++ apps/api/src/graphql/schema-discovery.ts
@@ -31,7 +31,7 @@ export function discoverySDL(): string {
     ciConflicts:     Int!
     relationsCreated: Int!
     relationsRemoved: Int!
-    durationMs:      Int
+    durationMs:      Float
     errorMessage:    String
     startedAt:       String!
     completedAt:     String
